backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
botocore==1.29.108
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.7
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-filter==23.1
django-import-export==3.2.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dynaconf==3.1.12
et-xmlfile==1.1.0
hvac==1.1.0
idna==3.4
importlib-metadata==6.1.0
jmespath==1.0.1
loguru==0.6.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.8
Pillow==9.4.0
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2022.7
urllib3==1.26.15
validators==0.20.0
win32-setctime==1.1.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
